Modified the compatible-language logic to be more IE-friendly
IE does not allow for language-only values in the preferred languages.
In order for things to work as expected, I modified the logic to act 
as if each language designation includes a language-only value as well,
as is recommended by the W3C here:
http://www.w3.org/International/questions/qa-lang-priorities#langtagdetail

This change breaks compliance with RFC 2616, but I like the results better.